Bit Size, Module (beginning/end)
I added a function to get the largest bit in a function then add it to the output string.  I also included a bool for if a REG occurs then we add clk and rst to the input list.
I had to create a new output file so we can write the first line [module circuit(a,b,c.....)] to the file.  Then deleted the "temp" file we were writing too.